+++ packages/suite/src/actions/wallet/__tests__/blockchainActions.test.ts
@@ -46,6 +46,7 @@ const getInitialState = (
         transactions: transactionsReducer(
             {
                 transactions: transactions || {},
+                phishing: {},
                 fetchStatusDetail: {},
             },
             action,