+++ packages/suite/src/actions/wallet/__tests__/transactionActions.test.ts
@@ -57,6 +57,7 @@ describe('Transaction Actions', () => {
                 [account1.key]: [getWalletTransaction()],
                 [account2.key]: [getWalletTransaction()],
             },
+            phishing: {},
             fetchStatusDetail: {},
         });