+++ packages/suite/src/storage/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 26.3.0
 
 - create `suiteSyncOwners` object store
+- create `phishing` object store for persisting user "marked as not scam" transaction IDs per account
 
 ## 26.2.0